Add BoardArea abstract class to remove code repetitions
