+++ src/main/java/it/unibo/jurassiko/model/territory/impl/TerritoryImpl.java
@@ -1,23 +1,14 @@
 package it.unibo.jurassiko.model.territory.impl;
 
-import java.util.HashSet;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import it.unibo.jurassiko.model.territory.api.Territory;
 
-public class TerritoryImpl implements Territory {
+public class TerritoryImpl extends AbstractBoardArea<Territory> implements Territory {
 
-    private String name;
     private String continent;
-    @JsonProperty("neighbours")
-    private Set<String> neighbourNames;
 
-    @JsonIgnore
-    private Set<Territory> neighbours;
     @JsonIgnore
     private int dinoAmount = 0;
 
@@ -26,36 +17,8 @@ public class TerritoryImpl implements Territory {
     }
 
     @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
     public String getContinent() {
         return continent;
-    }
-
-    @Override
-    public Set<String> getNeighbourNames() {
-        return Set.copyOf(neighbourNames);
-    }
-
-    @Override
-    public void setNeighbours(final Set<Territory> neighbours) {
-        this.neighbours = new HashSet<>(neighbours);
-    }
-
-    @Override
-    public Set<Territory> getNeighbours() {
-        return Set.copyOf(neighbours);
-    }
-
-    @Override
-    public boolean isNeighbour(final String territoryName) {
-        return this.neighbours.stream()
-                .map(t -> t.getName())
-                .collect(Collectors.toSet())
-                .contains(territoryName);
     }
 
     @Override
